fix: pre-push review batch — test-only export leak, public hook location, main-landmark test, MemberProfile 2-up regression
Six items from both pre-push reviewers:

- matchMediaStub.ts imported a devDependency (@testing-library/react) and
  shipped in the tarball (nothing resolves it, but dependency/license
  scanners would flag it). Renamed to matchMediaStub.testutil.ts and added
  "!src/**/*.testutil.ts" to package.json's files field — generalizes to
  the next test-only helper. Verified with npm pack --dry-run: gone.
- useBelowBreakpoint was exported straight out of _internal/, this repo's
  own "do not touch" signal, from a line parked mid-Sortable-block. Moved
  the hook to its own src/hooks/useBelowBreakpoint.ts (constants/types stay
  in _internal/collapse.ts), updated Rail/AppLayout/index.ts to import from
  there, moved its test cases to src/hooks/useBelowBreakpoint.test.ts.
  Confirmed src/structure.test.ts only scans src/components/*, so the new
  src/hooks/ directory doesn't trip the four-file rule. Gave it its own
  AGENTS.md entry (was a buried clause in AppLayout's sidebarOverlayBelow
  bullet).
- useBelowBreakpoint's breakpoint param is now optional
  (breakpoint?: CollapseBreakpoint) instead of a required
  `| undefined` union — it was a type error to call useBelowBreakpoint()
  even though "no breakpoint" is an explicitly supported, documented case.
- AppLayout.test.tsx's root test only checked `[class*="main"]`, which
  passes even if <main> regresses to <div>. Added a getByRole('main')
  assertion and retitled the mistitled test. AGENTS.md now documents that
  the children slot is a <main> landmark (nesting your own <main> inside
  double-lands it).
- AppLayout.tokens.scss's "consumer override still wins" comment
  overclaimed — true for a nested scope, not for another :root rule at the
  same specificity. Tightened.
- MemberProfile.tsx's two FormRows lost columns={2} in an earlier pass,
  falling back to FormRow's 16rem default minColumnWidth — inside a 440px
  drawer that's not enough room for 2-up, so First/Last name and
  Phone/Timezone stacked on desktop too. minColumnWidth="10rem" restores
  2-up at the drawer's ~408px desktop width while still reflowing to 1-up
  in the ~316px phone drawer. Measured both in-browser (not just computed
  via arithmetic): 1400px → grid-template-columns: 191px 191px (2 cols);
  380px → 306px (1 col).
