+++ packages/design-system/package.json
@@ -33,7 +33,8 @@
     "!src/**/*.spec.ts",
     "!src/**/*.spec.tsx",
     "!src/**/*.spec.js",
-    "!src/**/*.spec.jsx"
+    "!src/**/*.spec.jsx",
+    "!src/**/*.testutil.ts"
   ],
   "publishConfig": {
     "registry": "https://npm.pkg.github.com",
