+++ packages/design-system/src/components/_internal/collapse.test.ts
@@ -1,8 +1,6 @@
 import { readFileSync } from 'node:fs';
 import { resolve } from 'node:path';
-import { renderHook } from '@testing-library/react';
-import { COLLAPSE_BREAKPOINT_PX, COLLAPSE_BREAKPOINTS, useBelowBreakpoint } from './collapse';
-import { stubMatchMedia } from './matchMediaStub';
+import { COLLAPSE_BREAKPOINT_PX, COLLAPSE_BREAKPOINTS } from './collapse';
 
 /**
  * The collapse breakpoint scale is duplicated by necessity: container-query
@@ -40,48 +38,5 @@ describe('collapse breakpoints — TS and SCSS stay in sync', () => {
     expect(values).toEqual([...new Set(values)]);
     expect(COLLAPSE_BREAKPOINT_PX.sm).toBeLessThan(COLLAPSE_BREAKPOINT_PX.md);
     expect(COLLAPSE_BREAKPOINT_PX.md).toBeLessThan(COLLAPSE_BREAKPOINT_PX.lg);
-  });
-});
-
-describe('useBelowBreakpoint', () => {
-  const original = Object.getOwnPropertyDescriptor(window, 'matchMedia');
-
-  afterEach(() => {
-    if (original) Object.defineProperty(window, 'matchMedia', original);
-    else delete (window as { matchMedia?: unknown }).matchMedia;
-  });
-
-  it('returns false when no breakpoint is given', () => {
-    stubMatchMedia(320);
-    const { result } = renderHook(() => useBelowBreakpoint(undefined));
-    expect(result.current).toBe(false);
-  });
-
-  it('returns true below the threshold and false above it', () => {
-    stubMatchMedia(700);
-    const { result } = renderHook(() => useBelowBreakpoint('lg'));
-    expect(result.current).toBe(true);
-  });
-
-  it('is inclusive at the threshold value', () => {
-    stubMatchMedia(768);
-    const { result } = renderHook(() => useBelowBreakpoint('lg'));
-    expect(result.current).toBe(true);
-  });
-
-  it('updates when the viewport crosses the threshold', () => {
-    const mm = stubMatchMedia(1200);
-    const { result } = renderHook(() => useBelowBreakpoint('lg'));
-    expect(result.current).toBe(false);
-    mm.resizeTo(500);
-    expect(result.current).toBe(true);
-    mm.resizeTo(1000);
-    expect(result.current).toBe(false);
-  });
-
-  it('returns false when matchMedia is unavailable', () => {
-    delete (window as { matchMedia?: unknown }).matchMedia;
-    const { result } = renderHook(() => useBelowBreakpoint('sm'));
-    expect(result.current).toBe(false);
   });
 });
